requirements.txt
aiofile==3.7.4
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
azure-core==1.21.1
azure-identity==1.7.1
azure-quantum==0.22.187631
azure-storage-blob==12.9.0
caio==0.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Deprecated==1.2.13
dill==0.3.4
frozenlist==1.3.0
idna==3.3
isodate==0.6.1
mpmath==1.2.1
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
multidict==6.0.2
ntlm-auth==1.5.0
numpy==1.22.1
oauthlib==3.2.0
pbr==5.8.0
ply==3.11
portalocker==2.3.2
protobuf==3.19.4
psutil==5.9.0
pycparser==2.21
PyJWT==2.3.0
python-constraint==1.4.0
python-dateutil==2.8.2
qiskit==0.34.1
qiskit-aer==0.10.2
qiskit-ibmq-provider==0.18.3
qiskit-ignis==0.7.0
qiskit-ionq==0.1.4
qiskit-terra==0.19.1
requests==2.27.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
retworkx==0.11.0
scipy==1.7.3
six==1.16.0
stevedore==3.5.0
symengine==0.8.1
sympy==1.9
tweedledum==1.1.1
urllib3==1.26.8
websocket-client==1.2.3
wrapt==1.13.3
yarl==1.7.2
